extend: Make `parent` field unnecessary
Previously, the `parent` property of NamedBlocks is used as a way to
automatically update the node in `declaredBlocks` when the corresponding
node in the main AST is changed. While it might be more convenient, a
cleaner way is to update `declaredBlocks` after doing possible
replacement to the NamedBlocks. It is also somewhat necessary for the
next commit, as `parent` would contain circular references unsuitable
for JSON output used by the regression tests.